Bi-orthogonal Hartree-Fock

ElemCo.BOHF — Module

bi-orthogonal Hartree-Fock method (using a similarity-transformed FciDump)

Exported functions

ElemCo.BOHF.bohf — Method

bohf(EC::ECInfo)

Perform BO-HF using integrals from fcidump EC.fd.

ElemCo.BOHF.bouhf — Method

bouhf(EC::ECInfo)

Perform BO-UHF using integrals from fcidump EC.fd.

ElemCo.BOHF.guess_boorb — Function

guess_boorb(EC::ECInfo, guess::Symbol, uhf=false)

Calculate starting guess for BO-MO coefficients (left and right). Type of initial guess for MO coefficients is given by guess. uhf indicates whether the calculation is restricted or unrestricted.

See ScfOptions.guess for possible values. (Note: :SAD` is not possible here and will be replaced by identity matrix!)