ElemCo.jl methods

ElemCo.ECMethods — Module

Specify methods available for electron-correlation calculations

Main structure

ElemCo.ECMethods.ECMethod — Type

ECMethod

Description of the electron-correlation method

theory: theory level: "MP", "CC", "DC".

prefix: prefix of the methods, e.g., "EOM", "U", "R", "2D", "FRS", "FRT".

suffix: suffix of the methods.

exclevel: excitation level for each class (exclevel[1] for singles etc.). Possible values: :none, :full, :pert, :pertiter.

pertlevel: perturbation theory level (relevant for MP methods).

Exported functions

ElemCo.ECMethods.has_prefix — Method

has_prefix(method::ECMethod, spec::AbstractString)

Return true if method has prefix spec, e.g., "EOM".

ElemCo.ECMethods.has_suffix — Method

has_suffix(method::ECMethod, spec::AbstractString)

Return true if method has suffix spec, e.g., "F12".

ElemCo.ECMethods.is_unrestricted — Method

is_unrestricted(method::ECMethod)

Return true if method is unrestricted.

ElemCo.ECMethods.max_full_exc — Method

max_full_exc(method::ECMethod)

Return the highest full excitation level of the method.

ElemCo.ECMethods.method_name — Function

method_name(method::ECMethod, main::Bool = true)

Return string representation of method. If main is true, return only the main part of the name, i.e., without perturbative corrections.

ElemCo.ECMethods.set_prefix! — Method

set_prefix!(method::ECMethod, spec::AbstractString)

Set method to have prefix spec, e.g., "EOM".

ElemCo.ECMethods.set_suffix! — Method

set_suffix!(method::ECMethod, spec::AbstractString)

Set method to have suffix spec, e.g., "F12".

ElemCo.ECMethods.set_unrestricted! — Method

set_unrestricted!(method::ECMethod)

Set method to unrestricted.

Internal functions

Base.show — Method

show(io::IO, method::ECMethod)

Print method to io.

ElemCo.ECMethods.check_specs — Method

check_specs(mname::AbstractString, pos, specs::Vector)

Check if starting from position pos, mname contains any of specs and return a list of the matching ones (without dashes for multiple-letter specs!) and the final position after specs.