Molden Interface

ElemCo.MoldenInterface — Module

Molden interface

This module provides an interface to Molden to read and write orbitals and other data.

Exported functions and types

ElemCo.MoldenInterface.is_molden_file — Method

is_molden_file(filename::String)

Check if the file filename is a Molden file.

ElemCo.MoldenInterface.write_molden_orbitals — Method

write_molden_orbitals(EC::ECInfo, filename::String)

Write the current orbitals to a Molden file.

Internal functions and types

ElemCo.MoldenInterface.ao_permutation — Function

ao_permutation(EC::ECInfo, back=false)

Return the permutation of the atomic orbitals from the Molden to the libcint order such that μ(molden)[ao_permutation(EC)] = μ(libcint).

If back is true, the permutation is for the libcint to Molden order.

ElemCo.MoldenInterface.printmos — Function

printmos(f, orbs, order, energies, occupation, spin="Alpha")

Print the molecular orbital coefficients to a Molden file.

ElemCo.MoldenInterface.MOLDEN2LIBCINT_PERMUTATION — Constant

MOLDEN2LIBCINT_PERMUTATION

Permutation of the atomic orbitals from the Molden to the libcint order.

ElemCo.MoldenInterface.MOLDEN2LIBCINT_PERMUTATION_CART — Constant

MOLDEN2LIBCINT_PERMUTATION_CART

Permutation of the atomic orbitals from the Molden to the libcint order for cartesian basis sets.