Outputs

ElemCo.Outputs — Module

Outputs

Module for output functions.

The main purpose of this module is to hide the output functions from the JET.jl analyser.

Exported functions

ElemCo.Outputs.flush_output — Method

flush_output()

Flush the output buffer.

ElemCo.Outputs.output_E_method — Function

output_E_method(En, method::AbstractString, info::AbstractString="")

Output energy En with method method and additional info info.

ElemCo.Outputs.output_E_var — Method

output_E_var(En, var, Δt)

Output energy En, variance var, and time step Δt.

ElemCo.Outputs.output_iteration — Method

output_iteration(it, var, Δt, floats...)

Output iteration number it, variance var, time step Δt, and additional floats.

ElemCo.Outputs.output_norms — Method

output_norms(norms::Vector{Pair{String,Float64}})

Output norms.

The norms are a vector of pairs of strings and floats.

ElemCo.Outputs.output_time — Method

output_time(Δtime, info::AbstractString)

Output time with message info.

Internal functions