Calculations

ElemCo.ElemCo — Module

       ╭─────────────╮
Electron Correlation methods
       ╰─────────────╯

The ElemCo module contains the main macros and functions for running electronic structure calculations. The methods are contained in various submodules and are described in the following sections.

Various macros are defined and exported to simplify running calculations. The macros use several reserved variable names. The following table lists the reserved variable names and their meanings.

Variable	Meaning

EC::ECInfo	A global information object containing options, molecular system description, integrals and orbital spaces information, see ElemCo.ECInfo.

geometry::String	Molecular coordinates, either in the xyz format or the file containing the xyz coordinates, see ElemCo.MSystem.

basis::Dict	Basis set information, see ElemCo.MSystem

fcidump::String	File containing the integrals in the FCIDUMP format, see ElemCo.FciDump.

Macros

ElemCo.@ECinit — Macro

@ECinit()

Initialize EC::ECInfo and add molecular system and/or fcidump if variables geometry::String and basis::Dict{String,Any} and/or fcidump::String are defined.

If EC is already initialized, it will be overwritten.

Examples

geometry="He 0.0 0.0 0.0"
basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
@ECinit
# output
Occupied orbitals:[1]

ElemCo.@bohf — Macro

@bohf()

Run bi-orthogonal HF calculation using FCIDUMP integrals.

The orbitals are stored to WfOptions.orb. For open-shell systems (or UHF FCIDUMPs), the BO-UHF energy is calculated.

Examples

fcidump = "FCIDUMP"
@bohf

ElemCo.@bouhf — Macro

@bouhf()

Run bi-orthogonal UHF calculation using FCIDUMP integrals.

ElemCo.@cc — Macro

@cc(method, kwargs...)

Run coupled cluster calculation.

The type of the method is determined by the first argument (ccsd/ccsd(t)/dcsd etc). The method can be specified as a string or as a variable, e.g., @cc CCSD or @cc "CCSD" or ccmethod="CCSD"; @cc ccmethod.

Keyword arguments

fcidump::String: fcidump file (default: "", i.e., use integrals from EC).

occa::String: occupied α orbitals (default: "-").

occb::String: occupied β orbitals (default: "-").

Examples

geometry="bohr
O      0.000000000    0.000000000   -0.130186067
H1     0.000000000    1.489124508    1.033245507
H2     0.000000000   -1.489124508    1.033245507"
basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
@dfhf
@dfints
@cc ccsd

ElemCo.@copyfile — Macro

@copyfile(from_file, to_file, kwargs...)

Copy file from_file to to_file in EC.scr directory.

Keyword arguments

overwrite::Bool: overwrite existing file (default: false).

ElemCo.@dfhf — Macro

@dfhf()

Run DF-HF calculation. The orbitals are stored to WfOptions.orb.

ElemCo.@dfints — Macro

@dfints()

Generate 2 and 4-idx MO integrals using density fitting. The MO coefficients are read from WfOptions.orb.

ElemCo.@dfuhf — Macro

@dfuhf()

Run DF-UHF calculation. The orbitals are stored to WfOptions.orb.

ElemCo.@freeze_orbs — Macro

@freeze_orbs(freeze_orbs)

Freeze orbitals in the integrals according to an array or range freeze_orbs.

Examples

fcidump = "FCIDUMP"
@freeze_orbs 1:5
...
@ECinit
@freeze_orbs [1,2,20,21]

ElemCo.@loadfile — Macro

@loadfile(filename)

Read file filename from EC.scr directory.

Example

fock = @loadfile("f_mm")
orbs = @loadfile("C_Am")

ElemCo.@mainname — Macro

@mainname(file)

Return the main name of a file, i.e. the part before the last dot and the extension.

Examples

julia> @mainname("~/test.xyz")
("test", "xyz")

ElemCo.@method2string — Macro

@method2string(method, strmethod="")

Return string representation of method.

If method is a String variable, return the value of the variable. Otherwise, return the string representation of method (or strmethod if provided).

Examples

julia> @method2string(CCSD)
"CCSD"
julia> CCSD = "UCCSD";
julia> @method2string(CCSD)
"UCCSD"

ElemCo.@opt — Macro

@opt(what, kwargs...)

Set options for EC::ECInfo.

The first argument what is the name of the option (e.g., scf, cc, cholesky), see ECInfos.Options. The keyword arguments are the options to be set (e.g., thr=1.e-14, maxit=10). The current state of the options can be stored in a variable, e.g., opt_cc = @opt cc. The state can then be restored by @opt cc opt_cc. If EC is not already initialized, it will be done.

Examples

optscf = @opt scf thr=1.e-14 maxit=10
@opt cc maxit=100
...
@opt scf optscf

ElemCo.@reset — Macro

@reset(opt)

Reset options for opt to default values.

ElemCo.@run — Macro

general runner

ElemCo.@savefile — Macro

@savefile(filename, arr, kwargs...)

Save array or tuple of arrays arr to file filename in EC.scr directory.

Keyword arguments

description::String: description of the file (default: "tmp").

overwrite::Bool: overwrite existing file (default: false).

ElemCo.@svdcc — Macro

@svdcc(method="dcsd")

Run coupled cluster calculation with SVD decomposition of the amplitudes.

The type of the method is determined by the first argument (dcsd/dcd)

Examples

geometry="bohr
O      0.000000000    0.000000000   -0.130186067
H1     0.000000000    1.489124508    1.033245507
H2     0.000000000   -1.489124508    1.033245507"
basis = Dict("ao"=>"cc-pVDZ", "jkfit"=>"cc-pvtz-jkfit", "mp2fit"=>"cc-pvdz-rifit")
@dfhf
@svdcc

ElemCo.@transform_ints — Macro

@transform_ints(type="")

Rotate FCIDump integrals using WfOptions.orb as transformation matrices.

The orbitals are read from WfOptions.orb. If type is one of [bo, BO, bi-orthogonal, Bi-orthogonal, biorth, biorthogonal, Biorthogonal], the bi-orthogonal orbitals are used and the left transformation matrix is read from WfOptions.orb*WfOptions.left.

ElemCo.@tryECinit — Macro

@tryECinit()

If EC::ECInfo is not yet initialized, run @ECinit macro.

ElemCo.@write_ints — Macro

@write_ints(file="FCIDUMP", tol=-1.0)

Write FCIDump integrals to file file.

If tol is negative, all integrals are written, otherwise only integrals with absolute value larger than tol are written.

Exported functions

ElemCo.ECdriver — Method

ECdriver(EC::ECInfo, methods; fcidump="FCIDUMP", occa="-", occb="-")

Run electronic structure calculation for EC::ECInfo using methods methods::String.

The integrals are read from fcidump::String (default: "FCIDUMP"). If fcidump::String is empty, the integrals from EC.fd are used. The occupied α orbitals are given by occa::String (default: "-"). The occupied β orbitals are given by occb::String (default: "-"). If occb::String is empty, the occupied β orbitals are the same as the occupied α orbitals (closed-shell case).

Internal functions