Molpro Interface

ElemCo.MolproInterface — Module

MolproInterface

This module provides an interface to Molpro to read and write orbitals and other data.

Exported functions and types

ElemCo.MolproInterface.import_orbitals — Method

import_orbitals(EC::ECInfo, filename::AbstractString)

Import an orbital coefficient matrix from a Molpro matrop file.

ElemCo.MolproInterface.import_overlap — Method

import_overlap(EC::ECInfo, filename::AbstractString)

Import the overlap matrix from a Molpro matrop file.

ElemCo.MolproInterface.is_matrop_file — Method

is_matrop_file(filename::AbstractString)

Check if a file is a Molpro matrop file and return the type of the matrix.

ElemCo.MolproInterface.read_matrop_matrix — Method

read_matrop_matrix(filename::AbstractString)

Read a square matrix from a Molpro matrop file.

Internal functions and types

ElemCo.MolproInterface.ao_permutation — Method

ao_permutation(EC::ECInfo)

Return the permutation of the atomic orbitals from the Molpro to the libcint order such that μ(molpro)[ao_permutation(EC)] = μ(libcint).

ElemCo.MolproInterface.read_numbers_in_line — Method

read_numbers_in_line(f::IOStream)

Read a line from a file and return the numbers in it.

ElemCo.MolproInterface.skip_comment_lines — Method

skip_comment_lines(f::IOStream)

Skip lines which do not start with a number or a minus.

ElemCo.MolproInterface.MOLPRO2LIBCINT_PERMUTATION — Constant

MOLPRO2LIBCINT_PERMUTATION

Permutation of the atomic orbitals from the Molpro to the libcint order.