Options

ElemCo.ECInfos.Options — Type

Options for ElemCo.jl.

wf::ElemCo.ECInfos.WfOptions: Wavefunction options (WfOptions).

scf::ElemCo.ECInfos.ScfOptions: SCF options (ScfOptions).

int::ElemCo.ECInfos.IntOptions: Integral options (IntOptions).

cc::ElemCo.ECInfos.CcOptions: Coupled-Cluster options (CcOptions).

cholesky::ElemCo.ECInfos.CholeskyOptions: Cholesky options (CholeskyOptions).

diis::ElemCo.ECInfos.DiisOptions: DIIS options (DiisOptions).

ElemCo.ECInfos.CcOptions — Type

Options for Coupled-Cluster calculation.

thr::Float64: ⟨1.e-10⟩ convergence threshold.

maxit::Int64: ⟨50⟩ maximum number of iterations.

shifts::Float64: ⟨0.15⟩ level shift for singles.

shiftp::Float64: ⟨0.2⟩ level shift for doubles.

shiftt::Float64: ⟨0.2⟩ level shift for triples.

properties::Bool: ⟨false⟩ calculate properties.

ampsvdtol::Float64: ⟨1.e-3⟩ amplitude decomposition threshold.

ampsvdfac::Float64: ⟨1.e-2⟩ tightening amplitude decomposition factor (for the two-step decomposition).

use_kext::Bool: ⟨true⟩ use kext for doubles residual.

calc_d_vvvv::Bool: ⟨false⟩ calculate dressed <vv|vv>.

calc_d_vvvo::Bool: ⟨false⟩ calculate dressed <vv|vo>.

calc_d_vovv::Bool: ⟨false⟩ calculate dressed <vo|vv>.

calc_d_vvoo::Bool: ⟨false⟩ calculate dressed <vv|oo>.

triangular_kext::Bool: ⟨true⟩ use a triangular kext if possible.

calc_t3_for_decomposition::Bool: ⟨false⟩ calculate (T) for decomposition.

deco_ishiftp::Float64: ⟨0.0⟩ imaginary shift for denominator in doubles decomposition.

deco_ishiftt::Float64: ⟨0.0⟩ imaginary shift for denominator in triples decomposition.

use_projx::Bool: ⟨false⟩ use a projected exchange for contravariant doubles amplitudes in SVD-DCSD, $\tilde T_{XY} = U^{†a}_{iX} U^{†b}_{jY} \tilde T^{ij}_{ab}$.

use_full_t2::Bool: ⟨false⟩ use full doubles amplitudes in SVD-DCSD. The decomposition is used only for $N^6$ scaling terms.

project_vovo_t2::Int64: ⟨2⟩ what to project in $v_{ak}^{ci} T^{kj}_{cb}$ in SVD-DCSD: 0: both, 1: amplitudes, 2: residual, 3: robust fit.

decompose_full_doubles::Bool: ⟨false⟩ decompose full doubles amplitudes in SVD-DCSD (slow).

start::String: ⟨"cc_amplitudes"⟩ main part of filename for start amplitudes. For example, the singles amplitudes are read from start*"_singles"

save::String: ⟨"cc_amplitudes"⟩ main part of filename to save amplitudes. For example, the singles amplitudes are saved to save*"_singles"

start_lm::String: ⟨"cc_multipliers"⟩ main part of filename for start Lagrange multipliers. For example, the singles Lagrange multipliers are read from start_lm*"_singles"

save_lm::String: ⟨"cc_multipliers"⟩ main part of filename to save Lagrange multipliers. For example, the singles Lagrange multipliers are saved to save_lm*"_singles"

nomp2::Int64: ⟨0⟩ Don't use MP2 amplitudes as starting guess for the CC amplitudes

ElemCo.ECInfos.CholeskyOptions — Type

Options for Cholesky decomposition.

thr::Float64: ⟨1.e-6⟩ cholesky threshold.

ElemCo.ECInfos.DiisOptions — Type

Options for DIIS.

maxdiis::Int64: ⟨6⟩ maximum number of DIIS vectors.

resthr::Float64: ⟨10.0⟩ DIIS residual threshold.

ElemCo.ECInfos.IntOptions — Type

Options for integral calculation.

df::Bool: ⟨true⟩ use density-fitted integrals.

fcidump::String: ⟨""⟩ store integrals in FCIDump format.

ElemCo.ECInfos.ScfOptions — Type

Options for SCF calculation.

thr::Float64: ⟨1.e-10⟩ convergence threshold.

thren::Float64: ⟨sqrt(thr)*0.1⟩ energy convergence threshold (used additionally to thr).

maxit::Int64: ⟨50⟩ maximum number of iterations.

imagtol::Float64: ⟨1.e-8⟩ tolerance for imaginary part of MO coefs (for biorthogonal).

direct::Bool: ⟨false⟩ direct calculation without storing integrals.

guess::Symbol: ⟨:SAD⟩ orbital guess:

:HCORE from core Hamiltonian

:SAD from atomic densities

:GWH not implemented yet

:ORB from previous orbitals stored in file WfOptions.orb

temperature_guess::Float64: ⟨0.0⟩ Fermi-Dirac temperature for starting guess (at the moment works only for BO-HF).

ElemCo.ECInfos.WfOptions — Type

Options for wavefunction/orbitals.

ms2::Int64: ⟨-1⟩ spin magnetic quantum number times two (2×mₛ) of the system.

nelec::Int64: ⟨-1⟩ number of electrons. If < 0, the number of electrons is read from the FCIDump file or guessed for the neutral system.

charge::Int64: ⟨0⟩ charge of the system (relative to nelec/FCIDump/neutral system!).

orb::String: ⟨"C_Am"⟩ filename of MO coefficients. Used by all programs to read and write orbitals from/to file.

left::String: ⟨"-left"⟩ addition to the filename for left orbitals (for biorthogonal calculations).

core::Symbol: ⟨:large⟩ core type for frozen-core approximation:

:none no frozen-core approximation,

:small semi-core orbitals correlated,

:large semi-core orbitals frozen.

freeze_nocc::Int64: ⟨-1⟩ number of occupied (core) orbitals to freeze (overwrites core).

freeze_nvirt::Int64: ⟨0⟩ number of virtual (highest) orbitals to freeze.

occa::String: ⟨"-"⟩ occupied α (or closed-shell) orbitals.

occb::String: ⟨"-"⟩ occupied β orbitals. If occb::String is empty, the occupied β orbitals are the same as the occupied α orbitals (closed-shell case).